Fix failing tests and remove old device provisioning endpoint
