+++ kolibri/core/device/api_urls.py
@@ -5,7 +5,6 @@ from rest_framework import routers
 from .api import DeviceInfoView
 from .api import DeviceNameView
 from .api import DevicePermissionsViewSet
-from .api import DeviceProvisionView
 from .api import DeviceRestartView
 from .api import DeviceSettingsView
 from .api import DriveInfoViewSet
@@ -23,11 +22,6 @@ router.register(r"driveinfo", DriveInfoViewSet, basename="driveinfo")
 
 urlpatterns = [
     re_path(r"^", include(router.urls)),
-    re_path(
-        r"^deviceprovision/",
-        DeviceProvisionView.as_view({"post": "create"}),
-        name="deviceprovision",
-    ),
     re_path(r"^freespace/", FreeSpaceView.as_view({"get": "list"}), name="freespace"),
     re_path(r"^deviceinfo/", DeviceInfoView.as_view(), name="deviceinfo"),
     re_path(r"^devicesettings/", DeviceSettingsView.as_view(), name="devicesettings"),
